Air Dash, Stomp, Secondary Air Dash, Trigger Volume Ported
‡ Added stomp ability, which allows more decisive and precise platforming.

‡ Added an extra air dash which the player can perform if minifying in midair.

^ These two moves, alongside regular jumping and air dashing, can be chained.
‡ Ported interaction trigger volume into C++, inserting it between ASwitch and APressurePlate.

+ Learnt: Secondary air dash was not working at first. This was because a boolean && expression instantly succeeds if the first condition succeeeds. So I learnt I had to switch the most often true bool in the expression to second place.
+ Learnt: Originally, I used APawn::AddMovementInput on a timer for air dash, but this just made the player run without input. Whoops. I quickly changed it to ACharacter::LaunchCharacter (not on a timer) and it works fine now.

+++ Public/Switch.h
@@ -9,7 +9,8 @@
 
 // #Features to implement:
 // Timed: after a certain time, the switch returns to its original state
-// Once	: The switch can only be activated once (if (bIsActive && bExecuteOnce))
+// PRIVATE VARIABALES NEED TO BE GET- AND SETTABLE FROM WITHIN CHILD CLASSES: 
+// either just make them protected or add public/protected get- and setters.
 
 class USoundCue;
 
@@ -53,15 +54,15 @@ private:
 
 	// Whether this switch can only be used once
 	UPROPERTY(EditInstanceOnly, BlueprintReadOnly, Category = "Interaction", meta = (AllowPrivateAccess = "true"))
-	uint8 bExecuteOnce : 1;
+	uint32 bExecuteOnce : 1;
 
 	// Whether this switch can only be used once
 	UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category = "Interaction", meta = (AllowPrivateAccess = "true"))
-	uint8 bHasExecuted : 1;
+	uint32 bHasExecuted : 1;
 
 	// Whether this switch is turned on or off.
 	UPROPERTY(VisibleAnywhere, BlueprintReadWrite, Category = "Interaction", meta = (AllowPrivateAccess = "true"))
-	uint8 bIsActive : 1;
+	uint32 bIsActive : 1;
 
 	// Sound to play when the switch is turned on
 	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly, Category = "Interaction", meta = (AllowPrivateAccess = "true"))
